refactor: Align namespace
